Improve networking by changing the way input is handled
Previously it was a toggled state moving/not-moving. Now we send a
move-event every tick to server. Then server queues these events and
handles them together on every tick. Client-side prediction is still
here, but works differently. We now predict client state based on
acknowledged/not-acknowledged commands we sent to server.